add contiguous before g.view

+++ main.py
@@ -41,7 +41,7 @@ parser.add_argument('--log_dir', default='runs/pretrain', type=str)
 
 
 parser.add_argument('--optimizer', default='kfac', type=str)
-parser.add_argument('--batch_size', default=64, type=float)
+parser.add_argument('--batch_size', default=64, type=int)
 parser.add_argument('--epoch', default=100, type=int)
 parser.add_argument('--milestone', default=None, type=str)
 parser.add_argument('--learning_rate', default=1e-3, type=float) #0.01
